backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "HAHATON BROO"
authors = ["4got10king <Aldavletshina77654@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
jinja2 = "^3.1.2"
sqladmin = "0.16.1"
pydantic = { extras = ["email"], version = "^2.4.0" }
itsdangerous = "^2.1.2"
transliterate = "^1.10.2"
pyjwt = "^2.8.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
ultralytics = "^8.1.19"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
pytest = "^8.1.1"
aiosqlite = "^0.20.0"
python-logging-loki = "^0.3.1"
loguru = "^0.7.2"
loguru-loki-handler = "^0.1.1"
black = "^24.4.0"
faker = "^24.8.0"
pexpect = "^4.9.0"
starlette-exporter = "^0.21.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
aioboto3 = "^12.4.0"
bcrypt = "^4.1.3"
fastapi-pagination = "^0.12.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.9.0.post0"
alembic-postgresql-enum = "^1.3.0"
cyrtranslit = "^1.1.1"
pillow-heif = "^0.18.0"
aioftp = "^0.24.1"
aiofiles = "^24.1.0"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
